+++ src/lib/ngx-fileupload/libs/upload/src/upload.request.ts
@@ -1,7 +1,8 @@
 import { HttpClient, HttpEvent, HttpEventType, HttpProgressEvent, HttpResponse, HttpErrorResponse } from "@angular/common/http";
-import { Subject, Observable, forkJoin } from "rxjs";
+import { Subject, Observable, forkJoin, merge } from "rxjs";
 import { takeUntil, filter, switchMap, map } from "rxjs/operators";
-import { UploadState, UploadResponse, UploadData, Upload, Validator, ValidationFn} from "../../../data/api";
+import { UploadState, UploadResponse, UploadData, Upload} from "../../../data/api";
+import { Validator, ValidationFn } from "../../validation";
 import { UploadModel } from "../../../data/upload.model";
 
 /**
@@ -121,12 +122,13 @@ export class UploadRequest implements Upload {
     }
 
     public destroy() {
-        this.cancel$.next(true);
         this.destroyed$.next(true);
 
+        this.destroyed$.complete();
         this.cancel$.complete();
         this.upload$.complete();
 
+        this.destroyed$ = null;
         this.hooks      = null;
         this.upload$    = null;
         this.upload     = null;
@@ -200,6 +202,8 @@ export class UploadRequest implements Upload {
 
     /**
      * validate upload
+     *
+     * @deprecated
      */
     public validate(validator: Validator | ValidationFn) {
         const result = "validate" in validator
@@ -228,7 +232,7 @@ export class UploadRequest implements Upload {
         return this.http.post<string>(this.options.url, uploadBody, {
             reportProgress: true,
             observe: "events"
-        }).pipe(takeUntil(this.cancel$));
+        }).pipe(takeUntil(merge(this.cancel$, this.destroyed$)));
     }
 
     /**