+++ src/lib/ngx-fileupload/libs/validation/index.ts
@@ -1,2 +1,5 @@
 export * from "./src/validation.builder";
 export * from "./src/grouped.validator";
+export * from "./src/and.validator";
+export * from "./src/or.validator";
+export * from "./src/api";