+++ src/server/upload-server.js
@@ -4,7 +4,7 @@ const app = express();
 const resolve = require("path").resolve;
 const dirname = require("path").dirname;
 const letsLog = require("letslog");
-const fileUpload = require('express-fileupload');
+const fileUpload = require("express-fileupload");
 const logger = new letsLog.Logger({
     baseComment: "",
     loglvl: letsLog.ELoglevel.DEBUG,
@@ -25,13 +25,13 @@ const logger = new letsLog.Logger({
 let response = null;
 let timeout = 0;
 
-function sendResponse(res, file) {
+function sendResponse(res, file, msg = null) {
     const defaultResponse = {
         file: {
             id: 0,
-            type: 'any'
+            type: "any"
         },
-        message: `Hoooray File: ${file.name} uploaded to /dev/null`
+        message: msg || `Hoooray File: ${file.name} uploaded to /dev/null`
     };
 
     res.status(response ? response.state : 200);
@@ -48,12 +48,26 @@ app.use(fileUpload());
 app.post("/upload", function(req, res) {
 
     const uploadedFile = req.files.file;
-    logger.info(`File uploaded: ${uploadedFile.name}`)
+    logger.info(`File uploaded: ${uploadedFile.name}`);
 
     if(timeout) {
         setTimeout(() => sendResponse(res, uploadedFile), timeout);
     } else {
         sendResponse(res, uploadedFile);
+    }
+});
+
+app.post("/upload/gallery", function(req, res) {
+
+    const uploadedFile = req.files.picture;
+    logger.info(`Picture uploaded: ${uploadedFile.name}`);
+
+    const message = `New picture added to our gallery ${uploadedFile.name}`;
+
+    if(timeout) {
+        setTimeout(() => sendResponse(res, uploadedFile, message), timeout);
+    } else {
+        sendResponse(res, uploadedFile, message);
     }
 });